+++ packages/@headlessng-primitives/checkbox/src/lib/checkbox.directive.ts
@@ -15,12 +15,12 @@ import { ControlFieldRef } from '@headlessng/primitives/field';
 import { FocusDirective } from '@headlessng/primitives/focus';
 import { RequiredDirective } from '@headlessng/primitives/required';
 
-import { stateFromValue, CheckboxValue } from './checkbox.interface';
+import { stateFromChecked, CheckboxChecked } from './checkbox.interface';
 
 @Directive({
   exportAs: 'hCheckboxRef',
   host: {
-    '[attr.aria-checked]': 'value()',
+    '[attr.aria-checked]': '_checked()',
     '[attr.aria-describedby]': '_fieldRef?.descriptionId()',
     '[attr.aria-labelledby]': '_fieldRef?.labelId()',
     '[attr.data-state]': 'state()',
@@ -59,22 +59,22 @@ export class CheckboxDirective extends ControlFieldRef implements ControlValueAc
   public readonly focusRef = inject(FocusDirective);
   public readonly requiredRef = inject(RequiredDirective);
 
-  private readonly _value = signal<CheckboxValue>(false);
-  private readonly _valueEffect = effect(
+  private readonly _checked = signal<CheckboxChecked>(false);
+  private readonly _checkedEffect = effect(
     () => {
-      const value = this._value();
-      this._onChange?.(value);
+      const checked = this._checked();
+      this._onChange?.(checked);
       this._onTouched?.();
-      this.valueChange.emit(value);
+      this.checkedChange.emit(checked);
     },
     {
       injector: this._injector
     }
   );
 
-  public readonly state = computed(() => stateFromValue(this._value()));
-  public readonly value = this._value.asReadonly();
-  public readonly valueChange = output<CheckboxValue>();
+  public readonly checked = this._checked.asReadonly();
+  public readonly checkedChange = output<CheckboxChecked>();
+  public readonly state = computed(() => stateFromChecked(this._checked()));
 
   public readonly handleLabelClick = () => this._handleChange();
 
@@ -82,12 +82,12 @@ export class CheckboxDirective extends ControlFieldRef implements ControlValueAc
   @HostListener('click')
   private _handleChange(): void {
     if (!this.disabledRef.disabled()) {
-      this._value.set(!this._value());
+      this._checked.set(!this._checked());
     }
   }
 
-  private _onChange: ((value: CheckboxValue) => void) | undefined;
-  public registerOnChange(fn: (value: CheckboxValue) => void): void {
+  private _onChange: ((value: CheckboxChecked) => void) | undefined;
+  public registerOnChange(fn: (value: CheckboxChecked) => void): void {
     this._onChange = fn;
   }
 
@@ -100,7 +100,7 @@ export class CheckboxDirective extends ControlFieldRef implements ControlValueAc
     this.disabledRef.setDisabled(disabled);
   }
 
-  public writeValue(value: CheckboxValue): void {
-    this._value.set(value);
+  public writeValue(checked: CheckboxChecked): void {
+    this._checked.set(checked);
   }
 }